Move cake_day extension
